+++ packages/@webex/plugin-meetings/src/roap/request.ts
@@ -1,4 +1,3 @@
-/* global window */
 // @ts-ignore
 import {StatelessWebexPlugin} from '@webex/webex-core';